Refactored digits to hex_digits, count_hex_digits
This commit is the first stage of unifying the BigInt radix.
This introduces a new define BIGINT_RADIX which equals 16.
The algorithms and methods to support abritrary bases just
introduces an extra layer of difficulty and incompatibility.
As such support for arbitrary bases will be dropped and BigInt
will assume all inputs are in base 16, hexadecimal.