Implemented add and add_into, updated examples.
This commit provides the first feature set of addition and addition into
a BigInt structure. Note, currently addition does not take into account
the sign of the BigInt. This will be added in a future commit.
BigFibonacci example program was also updated to showcase how the BigInt
can be used.